+++ kernel/feature/kernel_umount.c
@@ -10,12 +10,12 @@
 #include <linux/printk.h>
 #include <linux/types.h>
 
-#include "kernel_umount.h"
+#include "feature/kernel_umount.h"
 #include "klog.h" // IWYU pragma: keep
-#include "allowlist.h"
+#include "policy/allowlist.h"
 #include "selinux/selinux.h"
-#include "feature.h"
-#include "ksud.h"
+#include "policy/feature.h"
+#include "runtime/ksud_boot.h"
 #include "ksu.h"
 
 static bool ksu_kernel_umount_enabled = true;
