+++ kernel/feature/sucompat.c
@@ -12,14 +12,14 @@
 #include <linux/sched/task_stack.h>
 #include <linux/ptrace.h>
 
-#include "allowlist.h"
-#include "feature.h"
+#include "policy/allowlist.h"
+#include "policy/feature.h"
 #include "klog.h" // IWYU pragma: keep
-#include "ksud.h"
-#include "sucompat.h"
-#include "app_profile.h"
+#include "runtime/ksud.h"
+#include "feature/sucompat.h"
+#include "policy/app_profile.h"
 
-extern void write_sulog(uint8_t sym);
+#include "tiny_sulog.h"
 
 #define SU_PATH "/system/bin/su"
 #define SH_PATH "/system/bin/sh"
