+++ kernel/hook/arm64/patch_memory.c
@@ -5,7 +5,7 @@
 #ifdef __aarch64__
 
 #include "../patch_memory.h"
-#include "../../klog.h" // IWYU pragma: keep
+#include "klog.h" // IWYU pragma: keep
 #include "linux/cpumask.h"
 #include "linux/gfp.h" // IWYU pragma: keep
 #include "linux/uaccess.h"
