+++ kernel/hook/syscall_event_bridge.c
@@ -1,0 +1,119 @@
+#include "linux/compiler.h"
+#include "linux/cred.h"
+#include "linux/jump_label.h"
+#include "linux/printk.h"
+#include "selinux/selinux.h"
+#include <asm/syscall.h>
+#include <linux/ptrace.h>
+#include <linux/static_key.h>
+
+#include "arch.h"
+#include "klog.h" // IWYU pragma: keep
+#include "hook/tp_marker.h"
+#include "feature/sucompat.h"
+#include "hook/setuid_hook.h"
+#include "policy/app_profile.h"
+#include "runtime/ksud.h"
+#include "hook/syscall_hook.h"
+#include "hook/syscall_event_bridge.h"
+
+static int ksu_handle_init_mark_tracker(const char __user **filename_user)
+{
+    char path[64];
+    unsigned long addr;
+    const char __user *fn;
+    long ret;
+
+    if (unlikely(!filename_user))
+        return 0;
+
+    addr = untagged_addr((unsigned long)*filename_user);
+    fn = (const char __user *)addr;
+    ret = strncpy_from_user(path, fn, sizeof(path));
+    if (ret < 0)
+        return 0;
+
+    path[sizeof(path) - 1] = '\0';
+    if (unlikely(strcmp(path, KSUD_PATH) == 0)) {
+        pr_info("hook_manager: escape to root for init executing ksud: %d\n", current->pid);
+        escape_to_root_for_init();
+    } else if (likely(strstr(path, "/app_process") == NULL && strstr(path, "/adbd") == NULL)) {
+        pr_info("hook_manager: unmark %d exec %s\n", current->pid, path);
+        ksu_clear_task_tracepoint_flag_if_needed(current);
+    }
+
+    return 0;
+}
+
+long __nocfi ksu_hook_newfstatat(int orig_nr, const struct pt_regs *regs)
+{
+    int *dfd;
+    const char __user **filename_user;
+    int *flags;
+
+    if (!ksu_su_compat_enabled)
+        return ksu_syscall_table[orig_nr](regs);
+
+    dfd = (int *)&PT_REGS_PARM1(regs);
+    filename_user = (const char __user **)&PT_REGS_PARM2(regs);
+    flags = (int *)&PT_REGS_SYSCALL_PARM4(regs);
+    ksu_handle_stat(dfd, filename_user, flags);
+
+    return ksu_syscall_table[orig_nr](regs);
+}
+
+long __nocfi ksu_hook_faccessat(int orig_nr, const struct pt_regs *regs)
+{
+    int *dfd;
+    const char __user **filename_user;
+    int *mode;
+
+    if (!ksu_su_compat_enabled)
+        return ksu_syscall_table[orig_nr](regs);
+
+    dfd = (int *)&PT_REGS_PARM1(regs);
+    filename_user = (const char __user **)&PT_REGS_PARM2(regs);
+    mode = (int *)&PT_REGS_PARM3(regs);
+    ksu_handle_faccessat(dfd, filename_user, mode, NULL);
+
+    return ksu_syscall_table[orig_nr](regs);
+}
+
+DEFINE_STATIC_KEY_TRUE(ksud_execve_key);
+
+void ksu_stop_ksud_execve_hook()
+{
+    static_branch_disable(&ksud_execve_key);
+}
+
+long __nocfi ksu_hook_execve(int orig_nr, const struct pt_regs *regs)
+{
+    int ret = 0;
+    const char __user **filename_user = (const char __user **)&PT_REGS_PARM1(regs);
+    bool current_is_init = is_init(current_cred());
+
+    if (static_branch_unlikely(&ksud_execve_key))
+        ksu_execve_hook_ksud(regs);
+
+    if (current->pid != 1 && current_is_init) {
+        ksu_handle_init_mark_tracker(filename_user);
+    } else if (ksu_su_compat_enabled) {
+        ret = ksu_handle_execve_sucompat(filename_user, NULL, NULL, NULL);
+        if (ret < 0)
+            return ret;
+    }
+
+    return ksu_syscall_table[orig_nr](regs);
+}
+
+long __nocfi ksu_hook_setresuid(int orig_nr, const struct pt_regs *regs)
+{
+    uid_t old_uid = current_uid().val;
+    long ret = ksu_syscall_table[orig_nr](regs);
+
+    if (ret < 0)
+        return ret;
+
+    ksu_handle_setresuid(old_uid, current_uid().val);
+    return ret;
+}
