+++ kernel/hook/syscall_hook_manager.c
@@ -1,8 +1,4 @@
-#include "linux/compiler.h"
-#include "linux/cred.h"
-#include "linux/jump_label.h"
 #include "linux/printk.h"
-#include "selinux/selinux.h"
 #include <linux/spinlock.h>
 #include <linux/kprobes.h>
 #include <linux/tracepoint.h>
@@ -10,23 +6,19 @@
 #include <linux/ptrace.h>
 #include <linux/slab.h>
 #include <trace/events/syscalls.h>
-#include <linux/static_key.h>
 
 #include "arch.h"
 #include "klog.h" // IWYU pragma: keep
-#include "syscall_hook_manager.h"
-#include "tp_marker.h"
-#include "sucompat.h"
-#include "setuid_hook.h"
-#include "selinux/selinux.h"
-#include "app_profile.h"
-#include "ksud.h"
+#include "hook/syscall_hook_manager.h"
+#include "hook/tp_marker.h"
+#include "feature/sucompat.h"
+#include "hook/setuid_hook.h"
 #include "hook/syscall_hook.h"
+#include "hook/syscall_event_bridge.h"
 
 #ifdef CONFIG_KRETPROBES
 
-static struct kretprobe *init_kretprobe(const char *name,
-                                        kretprobe_handler_t handler)
+static struct kretprobe *init_kretprobe(const char *name, kretprobe_handler_t handler)
 {
     struct kretprobe *rp = kzalloc(sizeof(struct kretprobe), GFP_KERNEL);
     if (!rp)
@@ -57,8 +49,7 @@ static void destroy_kretprobe(struct kretprobe **rp_ptr)
     *rp_ptr = NULL;
 }
 
-static int syscall_regfunc_handler(struct kretprobe_instance *ri,
-                                   struct pt_regs *regs)
+static int syscall_regfunc_handler(struct kretprobe_instance *ri, struct pt_regs *regs)
 {
     unsigned long flags;
     ksu_tp_marker_lock(&flags);
@@ -74,8 +65,7 @@ static int syscall_regfunc_handler(struct kretprobe_instance *ri,
     return 0;
 }
 
-static int syscall_unregfunc_handler(struct kretprobe_instance *ri,
-                                     struct pt_regs *regs)
+static int syscall_unregfunc_handler(struct kretprobe_instance *ri, struct pt_regs *regs)
 {
     unsigned long flags;
     ksu_tp_marker_lock(&flags);
@@ -95,112 +85,6 @@ static struct kretprobe *syscall_regfunc_rp = NULL;
 static struct kretprobe *syscall_unregfunc_rp = NULL;
 #endif
 
-// Unmark init's child that are not zygote, adbd or ksud
-int ksu_handle_init_mark_tracker(const char __user **filename_user)
-{
-    char path[64];
-    unsigned long addr;
-    const char __user *fn;
-
-    if (unlikely(!filename_user))
-        return 0;
-
-    addr = untagged_addr((unsigned long)*filename_user);
-    fn = (const char __user *)addr;
-
-    long ret = strncpy_from_user(path, fn, sizeof(path));
-
-    if (ret < 0) {
-        // unreadable path; keep mark to avoid wrongly unmarking zygote
-        return 0;
-    }
-    // Ensure NUL-termination when the string was truncated
-    path[sizeof(path) - 1] = '\0';
-
-    if (unlikely(strcmp(path, KSUD_PATH) == 0)) {
-        pr_info("hook_manager: escape to root for init executing ksud: %d\n",
-                current->pid);
-        escape_to_root_for_init();
-    } else if (likely(strstr(path, "/app_process") == NULL &&
-                      strstr(path, "/adbd") == NULL)) {
-        pr_info("hook_manager: unmark %d exec %s\n", current->pid, path);
-        ksu_clear_task_tracepoint_flag_if_needed(current);
-    }
-
-    return 0;
-}
-
-// Syscall hook handlers using the dispatcher signature
-static long __nocfi ksu_hook_newfstatat(int orig_nr, const struct pt_regs *regs)
-{
-    if (!ksu_su_compat_enabled)
-        return ksu_syscall_table[orig_nr](regs);
-
-    int *dfd = (int *)&PT_REGS_PARM1(regs);
-    const char __user **filename_user =
-        (const char __user **)&PT_REGS_PARM2(regs);
-    int *flags = (int *)&PT_REGS_SYSCALL_PARM4(regs);
-    ksu_handle_stat(dfd, filename_user, flags);
-
-    return ksu_syscall_table[orig_nr](regs);
-}
-
-static long __nocfi ksu_hook_faccessat(int orig_nr, const struct pt_regs *regs)
-{
-    if (!ksu_su_compat_enabled)
-        return ksu_syscall_table[orig_nr](regs);
-
-    int *dfd = (int *)&PT_REGS_PARM1(regs);
-    const char __user **filename_user =
-        (const char __user **)&PT_REGS_PARM2(regs);
-    int *mode = (int *)&PT_REGS_PARM3(regs);
-    ksu_handle_faccessat(dfd, filename_user, mode, NULL);
-
-    return ksu_syscall_table[orig_nr](regs);
-}
-
-DEFINE_STATIC_KEY_TRUE(ksud_execve_key);
-
-void ksu_stop_ksud_execve_hook()
-{
-    static_branch_disable(&ksud_execve_key);
-}
-
-static long __nocfi ksu_hook_execve(int orig_nr, const struct pt_regs *regs)
-{
-    int ret = 0;
-
-    const char __user **filename_user =
-        (const char __user **)&PT_REGS_PARM1(regs);
-    bool current_is_init = is_init(current_cred());
-    if (static_branch_unlikely(&ksud_execve_key))
-        ksu_execve_hook_ksud(regs);
-    if (current->pid != 1 && current_is_init) {
-        ksu_handle_init_mark_tracker(filename_user);
-    } else if (ksu_su_compat_enabled) {
-        ret = ksu_handle_execve_sucompat(filename_user, NULL, NULL, NULL);
-        if (ret < 0) {
-            return ret;
-        }
-    }
-
-    return ksu_syscall_table[orig_nr](regs);
-}
-
-static long __nocfi ksu_hook_setresuid(int orig_nr, const struct pt_regs *regs)
-{
-    uid_t old_uid = current_uid().val;
-
-    long ret = ksu_syscall_table[orig_nr](regs);
-
-    if (ret < 0)
-        return ret;
-
-    ksu_handle_setresuid(old_uid, current_uid().val);
-
-    return ret;
-}
-
 #ifdef CONFIG_HAVE_SYSCALL_TRACEPOINTS
 // sys_enter handler: redirect hooked syscalls to the dispatcher
 static void ksu_sys_enter_handler(void *data, struct pt_regs *regs, long id)
@@ -217,9 +101,9 @@ static void ksu_sys_enter_handler(void *data, struct pt_regs *regs, long id)
 
     if (ksu_has_syscall_hook(id)) {
         struct pt_regs *current_regs = task_pt_regs(current);
-        
+
 #if defined(__x86_64__)
-        // Stash the original syscall number in ax. 
+        // Stash the original syscall number in ax.
         // We use ax because it currently just holds -ENOSYS and is safe to overwrite.
         current_regs->ax = id;
         current_regs->orig_ax = ksu_dispatcher_nr;
@@ -237,10 +121,8 @@ void ksu_syscall_hook_manager_init(void)
     pr_info("hook_manager: ksu_hook_manager_init called\n");
 
 #ifdef CONFIG_KRETPROBES
-    syscall_regfunc_rp =
-        init_kretprobe("syscall_regfunc", syscall_regfunc_handler);
-    syscall_unregfunc_rp =
-        init_kretprobe("syscall_unregfunc", syscall_unregfunc_handler);
+    syscall_regfunc_rp = init_kretprobe("syscall_regfunc", syscall_regfunc_handler);
+    syscall_unregfunc_rp = init_kretprobe("syscall_unregfunc", syscall_unregfunc_handler);
 #endif
 
     // Register syscall hooks via dispatcher
@@ -255,8 +137,7 @@ void ksu_syscall_hook_manager_init(void)
     ksu_mark_running_process_locked();
 #endif
     if (ret) {
-        pr_err("hook_manager: failed to register sys_enter tracepoint: %d\n",
-               ret);
+        pr_err("hook_manager: failed to register sys_enter tracepoint: %d\n", ret);
     } else {
         pr_info("hook_manager: sys_enter tracepoint registered\n");
     }
