+++ kernel/hook/syscall_hook_manager.h
