+++ kernel/hook/tp_marker.c
@@ -1,4 +1,4 @@
-#include "tp_marker.h"
+#include "hook/tp_marker.h"
 
 #include "linux/cred.h"
 #include <linux/spinlock.h>
@@ -6,7 +6,7 @@
 #include <linux/sched/signal.h>
 #include <linux/sched/task.h>
 
-#include "allowlist.h"
+#include "policy/allowlist.h"
 #include "klog.h" // IWYU pragma: keep
 #include "selinux/selinux.h"
