+++ kernel/include/klog.h
