+++ kernel/include/ksu.h
@@ -2,15 +2,15 @@
 #define __KSU_H_KSU
 
 #include <linux/types.h>
-#include <linux/workqueue.h>
 #include <linux/cred.h>
+#include <linux/workqueue.h>
 
 #define KERNEL_SU_VERSION KSU_VERSION
 #define KERNEL_SU_VERSION_TAG KSU_VERSION_TAG
 
-#define EVENT_POST_FS_DATA 1
-#define EVENT_BOOT_COMPLETED 2
-#define EVENT_MODULE_MOUNTED 3
+extern struct cred *ksu_cred;
+extern bool ksu_late_loaded;
+extern bool allow_shell;
 
 static inline int startswith(char *s, char *prefix)
 {
@@ -25,8 +25,5 @@ static inline int endswith(const char *s, const char *t)
 		return 1;
 	return strcmp(s + slen - tlen, t);
 }
-
-extern struct cred *ksu_cred;
-extern bool ksu_late_loaded;
 
 #endif
