+++ kernel/infra/seccomp_cache.c
@@ -6,7 +6,7 @@
 #include <linux/filter.h>
 #include <linux/seccomp.h>
 #include "klog.h" // IWYU pragma: keep
-#include "seccomp_cache.h"
+#include "infra/seccomp_cache.h"
 
 struct action_cache {
 	DECLARE_BITMAP(allow_native, SECCOMP_ARCH_NATIVE_NR);
