+++ kernel/infra/su_mount_ns.c
@@ -18,7 +18,7 @@
 #include "arch.h"
 #include "klog.h" // IWYU pragma: keep
 #include "ksu.h"
-#include "su_mount_ns.h"
+#include "infra/su_mount_ns.h"
 
 extern int path_mount(const char *dev_name, struct path *path,
                       const char *type_page, unsigned long flags,
