+++ kernel/manager/apk_sign.h
