+++ kernel/manager/manager_observer.h
@@ -1,0 +1,7 @@
+#ifndef __KSU_H_MANAGER_OBSERVER
+#define __KSU_H_MANAGER_OBSERVER
+
+int ksu_observer_init(void);
+void ksu_observer_exit(void);
+
+#endif // __KSU_H_MANAGER_OBSERVER
