+++ kernel/manager/throne_tracker.c
@@ -7,11 +7,11 @@
 #include <linux/types.h>
 #include <linux/version.h>
 
-#include "allowlist.h"
-#include "apk_sign.h"
+#include "policy/allowlist.h"
+#include "manager/apk_sign.h"
 #include "klog.h" // IWYU pragma: keep
-#include "manager.h"
-#include "throne_tracker.h"
+#include "manager/manager_identity.h"
+#include "manager/throne_tracker.h"
 
 uid_t ksu_manager_appid = KSU_INVALID_APPID;
