+++ kernel/manager/throne_tracker.h
