+++ kernel/policy/feature.c
@@ -1,4 +1,4 @@
-#include "feature.h"
+#include "policy/feature.h"
 #include "klog.h" // IWYU pragma: keep
 
 #include <linux/mutex.h>
