+++ kernel/selinux/sepolicy.c
@@ -10,7 +10,7 @@
 #include <linux/version.h>
 
 #include "sepolicy.h"
-#include "../klog.h" // IWYU pragma: keep
+#include "klog.h" // IWYU pragma: keep
 #include "ss/symtab.h"
 
 #define KSU_SUPPORT_ADD_TYPE
