--- kernel/supercalls.c
@@ -1,1127 +1,0 @@
-#include <linux/anon_inodes.h>
-#include <linux/capability.h>
-#include <linux/cred.h>
-#include <linux/err.h>
-#include <linux/fdtable.h>
-#include <linux/file.h>
-#include <linux/fs.h>
-#include <linux/slab.h>
-#include <linux/kprobes.h>
-#include <linux/syscalls.h>
-#include <linux/task_work.h>
-#include <linux/uaccess.h>
-#include <linux/version.h>
-#include <linux/pid.h>
-#include <linux/utsname.h> // utsname() and uts_sem
-
-#include "supercalls.h"
-#include "arch.h"
-#include "allowlist.h"
-#include "feature.h"
-#include "klog.h" // IWYU pragma: keep
-#include "ksu.h"
-#include "ksud.h"
-#include "kernel_umount.h"
-#include "manager.h"
-#include "selinux/selinux.h"
-#include "file_wrapper.h"
-#include "tp_marker.h"
-
-#include "tiny_sulog.c"
-
-// Permission check functions
-bool only_manager(void)
-{
-	return is_manager();
-}
-
-bool only_root(void)
-{
-	return current_uid().val == 0;
-}
-
-bool manager_or_root(void)
-{
-	return current_uid().val == 0 || is_manager();
-}
-
-bool always_allow(void)
-{
-	return true; // No permission check
-}
-
-bool allowed_for_su(void)
-{
-    bool is_allowed =
-        is_manager() || ksu_is_allow_uid_for_current(current_uid().val);
-    return is_allowed;
-}
-
-static int do_grant_root(void __user *arg)
-{
-	// we already check uid above on allowed_for_su()
-
-    write_sulog('i'); // log ioctl escalation
-
-    pr_info("allow root for: %d\n", current_uid().val);
-    escape_with_root_profile();
-
-	return 0;
-}
-
-static uint32_t ksuver_override = 0;
-
-static int do_get_info(void __user *arg)
-{
-    struct ksu_get_info_cmd cmd = { .version = KERNEL_SU_VERSION, .flags = 0 };
-
-#ifdef MODULE
-	cmd.flags |= KSU_GET_INFO_FLAG_LKM;
-#endif
-
-	if (ksuver_override)
-	    cmd.version = ksuver_override;
-
-	if (is_manager()) {
-		cmd.flags |= KSU_GET_INFO_FLAG_MANAGER;
-	}
-	if (ksu_late_loaded) {
-		cmd.flags |= KSU_GET_INFO_FLAG_LATE_LOAD;
-	}
-	cmd.features = KSU_FEATURE_MAX;
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("get_version: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_report_event(void __user *arg)
-{
-	struct ksu_report_event_cmd cmd;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		return -EFAULT;
-	}
-
-	switch (cmd.event) {
-	case EVENT_POST_FS_DATA: {
-		static bool post_fs_data_lock = false;
-		if (!post_fs_data_lock) {
-			post_fs_data_lock = true;
-			if (ksu_late_loaded) {
-				pr_info("post-fs-data skipped (late load)\n");
-			} else {
-				pr_info("post-fs-data triggered\n");
-				on_post_fs_data();
-			}
-		}
-		break;
-	}
-	case EVENT_BOOT_COMPLETED: {
-		static bool boot_complete_lock = false;
-		if (!boot_complete_lock) {
-			boot_complete_lock = true;
-			if (ksu_late_loaded) {
-				pr_info("boot_complete skipped (late load)\n");
-			} else {
-				pr_info("boot_complete triggered\n");
-				on_boot_completed();
-			}
-		}
-		break;
-	}
-	case EVENT_MODULE_MOUNTED: {
-		pr_info("module mounted!\n");
-		on_module_mounted();
-		break;
-	}
-	default:
-		break;
-	}
-
-	return 0;
-}
-
-static int do_set_sepolicy(void __user *arg)
-{
-	struct ksu_set_sepolicy_cmd cmd;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		return -EFAULT;
-	}
-
-	return handle_sepolicy((void __user *)cmd.data, cmd.data_len);
-}
-
-static int do_check_safemode(void __user *arg)
-{
-	struct ksu_check_safemode_cmd cmd;
-
-	cmd.in_safe_mode = ksu_is_safe_mode();
-
-	if (cmd.in_safe_mode) {
-		pr_warn("safemode enabled!\n");
-	}
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("check_safemode: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_new_get_allow_list_common(void __user *arg, bool allow)
-{
-    struct ksu_new_get_allow_list_cmd cmd;
-    int *arr = NULL;
-    int err = 0;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		return -EFAULT;
-	}
-
-    if (cmd.count) {
-        arr = kmalloc(sizeof(int) * cmd.count, GFP_KERNEL);
-        if (!arr) {
-            return -ENOMEM;
-        }
-    }
-
-    bool success =
-        ksu_get_allow_list(arr, cmd.count, &cmd.count, &cmd.total_count, allow);
-
-    if (!success) {
-        err = -EFAULT;
-        goto out;
-    }
-
-    if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-        pr_err("new_get_allow_list: copy_to_user count failed\n");
-        err = -EFAULT;
-        goto out;
-    }
-
-    if (cmd.count &&
-        copy_to_user(&((struct ksu_new_get_allow_list_cmd *)arg)->uids, arr,
-                     sizeof(int) * cmd.count)) {
-        pr_err("new_get_allow_list: copy_to_user uids failed\n");
-        err = -EFAULT;
-    }
-
-out:
-    if (arr) {
-        kfree(arr);
-    }
-    return err;
-}
-
-static int do_new_get_deny_list(void __user *arg)
-{
-    return do_new_get_allow_list_common(arg, false);
-}
-
-static int do_new_get_allow_list(void __user *arg)
-{
-    return do_new_get_allow_list_common(arg, true);
-}
-
-static int do_get_allow_list_common(void __user *arg, bool allow)
-{
-    int *arr = NULL;
-    int err = 0;
-    u16 count;
-    u32 out_count;
-    static const u16 kSize = 128;
-
-    arr = kmalloc(sizeof(int) * kSize, GFP_KERNEL);
-    if (!arr) {
-        return -ENOMEM;
-    }
-
-    bool success = ksu_get_allow_list(arr, kSize, &count, NULL, allow);
-
-    if (!success) {
-        err = -EFAULT;
-        goto out;
-    }
-
-    out_count = count;
-
-    if (copy_to_user(arg + offsetof(struct ksu_get_allow_list_cmd, count),
-                     &out_count, sizeof(u32))) {
-        pr_err("get_allow_list: copy_to_user count failed\n");
-        err = -EFAULT;
-        goto out;
-    }
-
-    if (copy_to_user(arg, arr, sizeof(u32) * count)) {
-        pr_err("get_allow_list: copy_to_user uids failed\n");
-        err = -EFAULT;
-    }
-
-out:
-    if (arr) {
-        kfree(arr);
-    }
-    return err;
-}
-
-static int do_get_deny_list(void __user *arg)
-{
-    return do_get_allow_list_common(arg, false);
-}
-
-static int do_get_allow_list(void __user *arg)
-{
-    return do_get_allow_list_common(arg, true);
-}
-
-static int do_uid_granted_root(void __user *arg)
-{
-	struct ksu_uid_granted_root_cmd cmd;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		return -EFAULT;
-	}
-
-	cmd.granted = ksu_is_allow_uid_for_current(cmd.uid);
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("uid_granted_root: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_uid_should_umount(void __user *arg)
-{
-	struct ksu_uid_should_umount_cmd cmd;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		return -EFAULT;
-	}
-
-	cmd.should_umount = ksu_uid_should_umount(cmd.uid);
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("uid_should_umount: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_get_manager_appid(void __user *arg)
-{
-	struct ksu_get_manager_appid_cmd cmd;
-
-	cmd.appid = ksu_get_manager_appid();
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("get_manager_appid: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_get_app_profile(void __user *arg)
-{
-	struct ksu_get_app_profile_cmd cmd;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		pr_err("get_app_profile: copy_from_user failed\n");
-		return -EFAULT;
-	}
-
-	if (!ksu_get_app_profile(&cmd.profile)) {
-		return -ENOENT;
-	}
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("get_app_profile: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_set_app_profile(void __user *arg)
-{
-    struct ksu_set_app_profile_cmd cmd;
-    int ret;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		pr_err("set_app_profile: copy_from_user failed\n");
-		return -EFAULT;
-	}
-
-    ret = ksu_set_app_profile(&cmd.profile);
-    if (!ret) {
-        ksu_persistent_allow_list();
-        ksu_mark_running_process();
-    }
-    return ret;
-}
-
-static int do_get_feature(void __user *arg)
-{
-	struct ksu_get_feature_cmd cmd;
-	bool supported;
-	int ret;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		pr_err("get_feature: copy_from_user failed\n");
-		return -EFAULT;
-	}
-
-
-	ret = ksu_get_feature(cmd.feature_id, &cmd.value, &supported);
-	cmd.supported = supported ? 1 : 0;
-
-	if (ret && supported) {
-		pr_err("get_feature: failed for feature %u: %d\n", cmd.feature_id, ret);
-		return ret;
-	}
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("get_feature: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_set_feature(void __user *arg)
-{
-	struct ksu_set_feature_cmd cmd;
-	int ret;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		pr_err("set_feature: copy_from_user failed\n");
-		return -EFAULT;
-	}
-
-
-	ret = ksu_set_feature(cmd.feature_id, cmd.value);
-	if (ret) {
-		pr_err("set_feature: failed for feature %u: %d\n", cmd.feature_id, ret);
-		return ret;
-	}
-
-	return 0;
-}
-
-static int do_get_wrapper_fd(void __user *arg)
-{
-    if (!ksu_file_sid) {
-        return -EINVAL;
-    }
-
-    struct ksu_get_wrapper_fd_cmd cmd;
-    if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-        pr_err("get_wrapper_fd: copy_from_user failed\n");
-        return -EFAULT;
-    }
-
-    return ksu_install_file_wrapper(cmd.fd);
-}
-
-static int do_manage_mark(void __user *arg)
-{
-	struct ksu_manage_mark_cmd cmd;
-	int ret = 0;
-
-	if (copy_from_user(&cmd, arg, sizeof(cmd))) {
-		pr_err("manage_mark: copy_from_user failed\n");
-		return -EFAULT;
-	}
-
-	switch (cmd.operation) {
-	case KSU_MARK_GET: {
-		// Get task mark status
-		ret = ksu_get_task_mark(cmd.pid);
-		if (ret < 0) {
-			pr_err("manage_mark: get failed for pid %d: %d\n", cmd.pid, ret);
-			return ret;
-		}
-		cmd.result = (u32)ret;
-		break;
-	}
-	case KSU_MARK_MARK: {
-		if (cmd.pid == 0) {
-			ksu_mark_all_process();
-		} else {
-			ret = ksu_set_task_mark(cmd.pid, true);
-			if (ret < 0) {
-				pr_err("manage_mark: set_mark failed for pid %d: %d\n", cmd.pid,
-					ret);
-				return ret;
-			}
-		}
-		break;
-	}
-	case KSU_MARK_UNMARK: {
-		if (cmd.pid == 0) {
-			ksu_unmark_all_process();
-		} else {
-			ret = ksu_set_task_mark(cmd.pid, false);
-			if (ret < 0) {
-				pr_err("manage_mark: set_unmark failed for pid %d: %d\n",
-					cmd.pid, ret);
-				return ret;
-			}
-		}
-		break;
-	}
-	case KSU_MARK_REFRESH: {
-		ksu_mark_running_process();
-		pr_info("manage_mark: refreshed running processes\n");
-		break;
-	}
-	default: {
-		pr_err("manage_mark: invalid operation %u\n", cmd.operation);
-		return -EINVAL;
-	}
-	}
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("manage_mark: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_get_hook_mode(void __user *arg)
-{
-	struct ksu_get_hook_mode_cmd cmd = {0};
-
-	strscpy(cmd.mode, "Kprobes", sizeof(cmd.mode));
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("get_hook_mode: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_get_version_tag(void __user *arg)
-{
-	struct ksu_get_version_tag_cmd cmd = {0};
-
-	strscpy(cmd.tag, KERNEL_SU_VERSION_TAG, sizeof(cmd.tag));
-
-	if (copy_to_user(arg, &cmd, sizeof(cmd))) {
-		pr_err("get_version_tag: copy_to_user failed\n");
-		return -EFAULT;
-	}
-
-	return 0;
-}
-
-static int do_nuke_ext4_sysfs(void __user *arg)
-{
-    struct ksu_nuke_ext4_sysfs_cmd cmd;
-    char mnt[256];
-    long ret;
-
-    if (copy_from_user(&cmd, arg, sizeof(cmd)))
-        return -EFAULT;
-
-    if (!cmd.arg)
-        return -EINVAL;
-
-    memset(mnt, 0, sizeof(mnt));
-
-    ret = strncpy_from_user(mnt, cmd.arg, sizeof(mnt));
-    if (ret < 0) {
-        pr_err("nuke ext4 copy mnt failed: %ld\\n", ret);
-        return -EFAULT; // 或者 return ret;
-    }
-
-    if (ret == sizeof(mnt)) {
-        pr_err("nuke ext4 mnt path too long\\n");
-        return -ENAMETOOLONG;
-    }
-
-    pr_info("do_nuke_ext4_sysfs: %s\n", mnt);
-
-    return nuke_ext4_sysfs(mnt);
-}
-
-struct list_head mount_list = LIST_HEAD_INIT(mount_list);
-DECLARE_RWSEM(mount_list_lock);
-
-static int add_try_umount(void __user *arg)
-{
-    struct mount_entry *new_entry, *entry, *tmp;
-    struct ksu_add_try_umount_cmd cmd;
-    char buf[256] = { 0 };
-
-    if (copy_from_user(&cmd, arg, sizeof cmd))
-        return -EFAULT;
-
-    switch (cmd.mode) {
-    case KSU_UMOUNT_WIPE: {
-        struct mount_entry *entry, *tmp;
-        down_write(&mount_list_lock);
-        list_for_each_entry_safe (entry, tmp, &mount_list, list) {
-            pr_info("wipe_umount_list: removing entry: %s\n",
-                    entry->umountable);
-            list_del(&entry->list);
-            kfree(entry->umountable);
-            kfree(entry);
-        }
-        up_write(&mount_list_lock);
-
-        return 0;
-    }
-
-    case KSU_UMOUNT_ADD: {
-        long len = strncpy_from_user(buf, (const char __user *)cmd.arg, 256);
-        if (len <= 0)
-            return -EFAULT;
-
-        buf[sizeof(buf) - 1] = '\0';
-
-        new_entry = kzalloc(sizeof(*new_entry), GFP_KERNEL);
-        if (!new_entry)
-            return -ENOMEM;
-
-        new_entry->umountable = kstrdup(buf, GFP_KERNEL);
-        if (!new_entry->umountable) {
-            kfree(new_entry);
-            return -ENOMEM;
-        }
-
-        down_write(&mount_list_lock);
-
-        // disallow dupes
-        // if this gets too many, we can consider moving this whole task to a kthread
-        list_for_each_entry (entry, &mount_list, list) {
-            if (!strcmp(entry->umountable, buf)) {
-                pr_info("cmd_add_try_umount: %s is already here!\n", buf);
-                up_write(&mount_list_lock);
-                kfree(new_entry->umountable);
-                kfree(new_entry);
-                return -EEXIST;
-            }
-        }
-
-        // now check flags and add
-        // this also serves as a null check
-        if (cmd.flags)
-            new_entry->flags = cmd.flags;
-        else
-            new_entry->flags = 0;
-
-        // debug
-        list_add(&new_entry->list, &mount_list);
-        up_write(&mount_list_lock);
-        pr_info("cmd_add_try_umount: %s added!\n", buf);
-
-        return 0;
-    }
-
-    // this is just strcmp'd wipe anyway
-    case KSU_UMOUNT_DEL: {
-        long len = strncpy_from_user(buf, (const char __user *)cmd.arg,
-                                     sizeof(buf) - 1);
-        if (len <= 0)
-            return -EFAULT;
-
-        buf[sizeof(buf) - 1] = '\0';
-
-        down_write(&mount_list_lock);
-        list_for_each_entry_safe (entry, tmp, &mount_list, list) {
-            if (!strcmp(entry->umountable, buf)) {
-                pr_info("cmd_add_try_umount: entry removed: %s\n",
-                        entry->umountable);
-                list_del(&entry->list);
-                kfree(entry->umountable);
-                kfree(entry);
-            }
-        }
-        up_write(&mount_list_lock);
-
-        return 0;
-    }
-
-    // this way userspace can deduce the memory it has to prepare.
-    case KSU_UMOUNT_GETSIZE: {
-        // check for pointer first
-        if (!cmd.arg)
-            return -EFAULT;
-        
-        size_t total_size = 0; // size of list in bytes
-
-        down_read(&mount_list_lock);
-        list_for_each_entry(entry, &mount_list, list) {
-            total_size = total_size + strlen(entry->umountable) + 1; // + 1 for \0
-        }
-        up_read(&mount_list_lock);
-
-        // debug
-        // pr_info("cmd_add_try_umount: total_size: %zu\n", total_size);
-            
-        if (copy_to_user((size_t __user *)cmd.arg, &total_size, sizeof(total_size)))
-            return -EFAULT;
-
-        return 0;
-    }
-        
-    // WARNING! this is straight up pointerwalking.
-    // this way we dont need to redefine the ioctl defs.
-    // this also avoids us needing to kmalloc
-    // userspace have to send pointer to memory (malloc/alloca) or pointer to a VLA.
-    case KSU_UMOUNT_GETLIST: {
-        // check for pointer first
-        if (!cmd.arg)
-            return -EFAULT;
-            
-        char *user_buf = (char *)cmd.arg;
-
-        down_read(&mount_list_lock);
-        list_for_each_entry(entry, &mount_list, list) {
-
-            //debug
-            //pr_info("cmd_add_try_umount: entry: %s\n", entry->umountable);
-            
-            if (copy_to_user((char __user *)user_buf, entry->umountable, strlen(entry->umountable) + 1 )) {
-                up_read(&mount_list_lock);
-                return -EFAULT;
-            }
-
-            // walk it! +1 for null terminator
-            user_buf = user_buf + strlen(entry->umountable) + 1;
-        }
-        up_read(&mount_list_lock);
-
-        return 0;
-    }
-
-    default: {
-        pr_err("cmd_add_try_umount: invalid operation %u\n", cmd.mode);
-        return -EINVAL;
-    }
-
-    } // switch(cmd.mode)
-
-    return 0;
-}
-
-static int do_set_init_pgrp(void __user *arg)
-{
-    int err;
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(6, 15, 0)
-    struct pid *pids[PIDTYPE_MAX] = { 0 };
-#endif
-
-    write_lock_irq(&tasklist_lock);
-    struct task_struct *p = current->group_leader;
-    struct pid *init_group = task_pgrp(&init_task);
-
-    err = -EPERM;
-    if (task_session(p) != task_session(&init_task))
-        goto out;
-
-    err = 0;
-    if (task_pgrp(p) != init_group) {
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(6, 15, 0)
-        change_pid(pids, p, PIDTYPE_PGID, init_group);
-#else
-        change_pid(p, PIDTYPE_PGID, init_group);
-#endif
-    }
-
-out:
-    write_unlock_irq(&tasklist_lock);
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(6, 15, 0)
-    free_pids(pids);
-#endif
-
-    return err;
-}
-
-// IOCTL handlers mapping table
-static const struct ksu_ioctl_cmd_map ksu_ioctl_handlers[] = {
-    { .cmd = KSU_IOCTL_GRANT_ROOT,
-      .name = "GRANT_ROOT",
-      .handler = do_grant_root,
-      .perm_check = allowed_for_su },
-    { .cmd = KSU_IOCTL_GET_INFO,
-      .name = "GET_INFO",
-      .handler = do_get_info,
-      .perm_check = always_allow },
-    { .cmd = KSU_IOCTL_REPORT_EVENT,
-      .name = "REPORT_EVENT",
-      .handler = do_report_event,
-      .perm_check = only_root },
-    { .cmd = KSU_IOCTL_SET_SEPOLICY,
-      .name = "SET_SEPOLICY",
-      .handler = do_set_sepolicy,
-      .perm_check = only_root },
-    { .cmd = KSU_IOCTL_CHECK_SAFEMODE,
-      .name = "CHECK_SAFEMODE",
-      .handler = do_check_safemode,
-      .perm_check = always_allow },
-    { .cmd = KSU_IOCTL_GET_ALLOW_LIST,
-      .name = "GET_ALLOW_LIST",
-      .handler = do_get_allow_list,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_GET_DENY_LIST,
-      .name = "GET_DENY_LIST",
-      .handler = do_get_deny_list,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_NEW_GET_ALLOW_LIST,
-      .name = "NEW_GET_ALLOW_LIST",
-      .handler = do_new_get_allow_list,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_NEW_GET_DENY_LIST,
-      .name = "NEW_GET_DENY_LIST",
-      .handler = do_new_get_deny_list,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_UID_GRANTED_ROOT,
-      .name = "UID_GRANTED_ROOT",
-      .handler = do_uid_granted_root,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_UID_SHOULD_UMOUNT,
-      .name = "UID_SHOULD_UMOUNT",
-      .handler = do_uid_should_umount,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_GET_MANAGER_APPID,
-      .name = "GET_MANAGER_APPID",
-      .handler = do_get_manager_appid,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_GET_APP_PROFILE,
-      .name = "GET_APP_PROFILE",
-      .handler = do_get_app_profile,
-      .perm_check = only_manager },
-    { .cmd = KSU_IOCTL_SET_APP_PROFILE,
-      .name = "SET_APP_PROFILE",
-      .handler = do_set_app_profile,
-      .perm_check = only_manager },
-    { .cmd = KSU_IOCTL_GET_FEATURE,
-      .name = "GET_FEATURE",
-      .handler = do_get_feature,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_SET_FEATURE,
-      .name = "SET_FEATURE",
-      .handler = do_set_feature,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_GET_WRAPPER_FD,
-      .name = "GET_WRAPPER_FD",
-      .handler = do_get_wrapper_fd,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_MANAGE_MARK,
-      .name = "MANAGE_MARK",
-      .handler = do_manage_mark,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_NUKE_EXT4_SYSFS,
-      .name = "NUKE_EXT4_SYSFS",
-      .handler = do_nuke_ext4_sysfs,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_ADD_TRY_UMOUNT,
-      .name = "ADD_TRY_UMOUNT",
-      .handler = add_try_umount,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_SET_INIT_PGRP,
-      .name = "SET_INIT_PGRP",
-      .handler = do_set_init_pgrp,
-      .perm_check = only_root },
-    { .cmd = KSU_IOCTL_GET_HOOK_MODE,
-      .name = "GET_HOOK_MODE",
-      .handler = do_get_hook_mode,
-      .perm_check = manager_or_root },
-    { .cmd = KSU_IOCTL_GET_VERSION_TAG,
-      .name = "GET_VERSION_TAG",
-      .handler = do_get_version_tag,
-      .perm_check = manager_or_root },
-    { .cmd = 0, .name = NULL, .handler = NULL, .perm_check = NULL } // Sentinel
-};
-
-struct ksu_install_fd_tw {
-	struct callback_head cb;
-	int __user *outp;
-};
-
-static void ksu_install_fd_tw_func(struct callback_head *cb)
-{
-    struct ksu_install_fd_tw *tw =
-        container_of(cb, struct ksu_install_fd_tw, cb);
-    int fd = ksu_install_fd();
-    pr_info("[%d] install ksu fd: %d\n", current->pid, fd);
-
-	if (copy_to_user(tw->outp, &fd, sizeof(fd))) {
-		pr_err("install ksu fd reply err\n");
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(5, 11, 0)
-		close_fd(fd);
-#else
-		ksys_close(fd);
-#endif
-	}
-
-	kfree(tw);
-}
-
-static int reboot_handler_pre(struct kprobe *p, struct pt_regs *regs)
-{
-	struct pt_regs *real_regs = PT_REAL_REGS(regs);
-	int magic1 = (int)PT_REGS_PARM1(real_regs);
-	int magic2 = (int)PT_REGS_PARM2(real_regs);
-	unsigned int cmd = (unsigned int)PT_REGS_PARM3(real_regs);
-	unsigned long arg4 = (unsigned long)PT_REGS_SYSCALL_PARM4(real_regs);
-	unsigned long reply = (unsigned long)arg4;
-
-	/* Check if this is a request to install KSU fd */
-	if (magic1 == KSU_INSTALL_MAGIC1 && magic2 == KSU_INSTALL_MAGIC2) {
-		struct ksu_install_fd_tw *tw;
-
-		tw = kzalloc(sizeof(*tw), GFP_ATOMIC);
-		if (!tw)
-			return 0;
-
-		tw->outp = (int __user *)arg4;
-		tw->cb.func = ksu_install_fd_tw_func;
-
-		if (task_work_add(current, &tw->cb, TWA_RESUME)) {
-			kfree(tw);
-			pr_warn("install fd add task_work failed\n");
-		}
-	}
-
-	if (magic2 == CHANGE_MANAGER_UID) {
-		/* only root is allowed for this command */
-		if (current_uid().val != 0)
-			return 0;
-
-		pr_info("sys_reboot: ksu_set_manager_appid to: %d\n", cmd);
-		ksu_set_manager_appid(cmd);
-
-		if (cmd == ksu_get_manager_appid()) {
-			if (copy_to_user((void __user *)arg4, &reply, sizeof(reply)))
-				pr_info("sys_reboot: reply fail\n");
-		}
-
-		return 0;
-	}
-
-	if (magic2 == GET_SULOG_DUMP_V2) {
-		if (current_uid().val != 0)
-			return 0;
-
-		int ret = send_sulog_dump((void __user *)arg4);
-            if (ret)
-                return 0;
-
-        if (copy_to_user((void __user *)arg4, &reply, sizeof(reply) ))
-            return 0;
-	}
-
-    if (magic2 == CHANGE_KSUVER) {
-        if (current_uid().val != 0)
-            return 0;
-
-        pr_info("sys_reboot: ksu_change_ksuver to: %d\n", cmd);
-        ksuver_override = cmd;
-
-        if (copy_to_user((void __user *)arg4, &reply, sizeof(reply) ))
-            return 0;
-    }
-
-	// WARNING!!! triple ptr zone! ***
-	// https://wiki.c2.com/?ThreeStarProgrammer
-	if (magic2 == CHANGE_SPOOF_UNAME) {
-		// only root is allowed for this command 
-		if (current_uid().val != 0)
-			return 0;
-
-		char release_buf[65];
-		char version_buf[65];
-		static char original_release_buf[65] = {0};
-		static char original_version_buf[65] = {0};
-
-		// basically void * void __user * void __user *arg
-		void __user **ppptr = (void __user **)arg4;
-
-		// user pointer storage
-		// init this as zero so this works on 32-on-64 compat (LE)
-		uint64_t u_pptr = 0;
-		uint64_t u_ptr = 0;
-
-		pr_info("sys_reboot: ppptr: 0x%lx \n", (uintptr_t)ppptr);
-
-		// arg here is ***, pull out user-space ** via copy_from_user
-		if (copy_from_user(&u_pptr, ppptr, sizeof(u_pptr)))
-			return 0;
-
-		pr_info("sys_reboot: u_pptr: 0x%lx \n", (uintptr_t)u_pptr);
-
-		// now we got the __user **
-		// we cannot dereference this as this is __user
-		// we just do another copy_from_user to get it
-		if (copy_from_user(&u_ptr, (void __user *)u_pptr, sizeof(u_ptr)))
-			return 0;
-
-		pr_info("sys_reboot: u_ptr: 0x%lx \n", (uintptr_t)u_ptr);
-
-		// for release
-		if (strncpy_from_user(release_buf, (char __user *)u_ptr, sizeof(release_buf)) < 0)
-			return 0;
-		release_buf[sizeof(release_buf) - 1] = '\0'; 
-
-		// for version
-		if (strncpy_from_user(version_buf, (char __user *)(u_ptr + strlen(release_buf) + 1), sizeof(version_buf)) < 0)
-			return 0;
-		version_buf[sizeof(version_buf) - 1] = '\0'; 
-
-		if (original_release_buf[0] == '\0') {
-			struct new_utsname *u_curr = utsname();
-			// we save current version as the original before modifying
-			strncpy(original_release_buf, u_curr->release, sizeof(original_release_buf));
-			strncpy(original_version_buf, u_curr->version, sizeof(original_version_buf));
-			pr_info("sys_reboot: original uname saved: %s %s\n", original_release_buf, original_version_buf);
-		}
-
-		// so user can reset
-		if (!strcmp(release_buf, "default") || !strcmp(version_buf, "default") ) {
-			memcpy(release_buf, original_release_buf, sizeof(release_buf));
-			memcpy(version_buf, original_version_buf, sizeof(version_buf));
-		}
-
-		pr_info("sys_reboot: spoofing kernel to: %s - %s\n", release_buf, version_buf);
-
-		struct new_utsname *u = utsname();
-
-		down_write(&uts_sem);
-		strncpy(u->release, release_buf, sizeof(u->release));
-		strncpy(u->version, version_buf, sizeof(u->version));
-		up_write(&uts_sem);
-
-		// we write our confirmation on **
-		if (copy_to_user((void __user *)arg4, &reply, sizeof(reply)))
-			return 0;
-	}
-
-	return 0;
-}
-
-static struct kprobe reboot_kp = {
-	.symbol_name = REBOOT_SYMBOL,
-	.pre_handler = reboot_handler_pre,
-};
-
-void ksu_supercalls_init(void)
-{
-	int i;
-
-    pr_info("KernelSU IOCTL Commands:\n");
-    for (i = 0; ksu_ioctl_handlers[i].handler; i++) {
-        pr_info("  %-18s = 0x%08x\n", ksu_ioctl_handlers[i].name,
-                ksu_ioctl_handlers[i].cmd);
-    }
-
-	int rc = register_kprobe(&reboot_kp);
-	if (rc) {
-		pr_err("reboot kprobe failed: %d\n", rc);
-	} else {
-		pr_info("reboot kprobe registered successfully\n");
-	}
-
-    sulog_init_heap(); // grab heap memory
-}
-
-void ksu_supercalls_exit(void)
-{
-    struct mount_entry *entry, *tmp;
-
-    unregister_kprobe(&reboot_kp);
-
-    down_write(&mount_list_lock);
-    list_for_each_entry_safe (entry, tmp, &mount_list, list) {
-        list_del(&entry->list);
-        kfree(entry->umountable);
-        kfree(entry);
-    }
-    up_write(&mount_list_lock);
-}
-
-// IOCTL dispatcher
-static long anon_ksu_ioctl(struct file *filp, unsigned int cmd,
-                           unsigned long arg)
-{
-	void __user *argp = (void __user *)arg;
-	int i;
-
-#ifdef CONFIG_KSU_DEBUG
-	pr_info("ksu ioctl: cmd=0x%x from uid=%d\n", cmd, current_uid().val);
-#endif
-
-    for (i = 0; ksu_ioctl_handlers[i].handler; i++) {
-        if (cmd == ksu_ioctl_handlers[i].cmd) {
-            // Check permission first
-            if (ksu_ioctl_handlers[i].perm_check &&
-                !ksu_ioctl_handlers[i].perm_check()) {
-                pr_warn("ksu ioctl: permission denied for cmd=0x%x uid=%d\n",
-                        cmd, current_uid().val);
-                return -EPERM;
-            }
-            // Execute handler
-            return ksu_ioctl_handlers[i].handler(argp);
-        }
-    }
-
-	pr_warn("ksu ioctl: unsupported command 0x%x\n", cmd);
-	return -ENOTTY;
-}
-
-// File release handler
-static int anon_ksu_release(struct inode *inode, struct file *filp)
-{
-	pr_info("ksu fd released\n");
-	return 0;
-}
-
-// File operations structure
-static const struct file_operations anon_ksu_fops = {
-	.owner = THIS_MODULE,
-	.unlocked_ioctl = anon_ksu_ioctl,
-	.compat_ioctl = anon_ksu_ioctl,
-	.release = anon_ksu_release,
-};
-
-// Install KSU fd to current process
-int ksu_install_fd(void)
-{
-	struct file *filp;
-	int fd;
-
-	// Get unused fd
-	fd = get_unused_fd_flags(O_CLOEXEC);
-	if (fd < 0) {
-		pr_err("ksu_install_fd: failed to get unused fd\n");
-		return fd;
-	}
-
-    // Create anonymous inode file
-    filp = anon_inode_getfile("[ksu_driver]", &anon_ksu_fops, NULL,
-                              O_RDWR | O_CLOEXEC);
-    if (IS_ERR(filp)) {
-        pr_err("ksu_install_fd: failed to create anon inode file\n");
-        put_unused_fd(fd);
-        return PTR_ERR(filp);
-    }
-
-	// Install fd
-	fd_install(fd, filp);
-
-	pr_info("ksu fd installed: %d for pid %d\n", fd, current->pid);
-
-	return fd;
-}
